revamp debugging infrastructure
To be easier to identify what each debug is about we're now using more
informative debugging messages and while in that we've also improved
the code to not require the duplicated parentsis.

During the process some debugging calls has been changes to be easier
to spot where they are comming from.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>

git-svn-id: https://freerdp.svn.sourceforge.net/svnroot/freerdp/trunk@182 7ffbcff5-ea86-4e99-a76d-b9f7f6476058