Handle Server Channel ID
This makes some difference for the VirtualBox RDP server which uses channel 0 instead of 0x03EA.

http://msdn.microsoft.com/en-us/library/cc240906%28v=PROT.10%29.aspx defines Server Channel ID

http://msdn.microsoft.com/en-us/library/cc240669%28PROT.10%29.aspx says about Processing Demand Active PDU:
the server MCS channel ID (present in the pduSource field) MUST be stored in the Server Channel ID store.

http://msdn.microsoft.com/en-us/library/cc240488%28PROT.10%29.aspx says about Confirm Active PDU Data:
originatorId ... This field MUST be set to the server channel ID (0x03EA).

http://msdn.microsoft.com/en-us/library/cc240671%28PROT.10%29.aspx says about Sending Synchronize PDU:
The targetUser field SHOULD be set to the MCS server channel ID (held in the Server Channel ID store)