bug fix: set stream pointer correctly after processing (un)compressed data
This also allows processing of several data PDUs inside the encrypted body of
one sendDataIndication userData.

AFAIK MS RDP servers never does that, and AFAICS the RDP specification do not
allow that. (Something similar is however permitted for Fast Path.)

Both mstsc and rdesktop do however process multiple data PDUs, and VirtualBox
4.0.0_69151 utilizes that.

We ignored the second PDU, and the assertions for correct processing of
compressed data noticed that.

Now we set the stream pointer correctly after the first PDU even if we didn't
process it. The assertions that we process the PDUs that we process correctly
has been updated too.