libfreerdp / xfreerdp: consistently use stdout for debug info
Previously some debug/error info was written buffered and unflushed on stdout and some
on stderr. The result was that debug output often wasn't complete or didn't
appear in the right order, especially (not) when redirecting to a file.

freerdp applications will by nature not be command line applications but GUI
applications that don't have any "ordinary" output on stdout. All output will
be for debugging, so consistently using stdout would be fine.