drop DEBUG_RDP5 and introduce DEBUG_DRAW
The distinction between "rdp5" and the rest of the protocol is artificial and
is not useful for debugging.

It is much more relevant to distinguish the drawing messages from the protocol
level debugging.

There is some overlap between these definitions, but now it is clarified that
it is DEBUG_DRAW for debugging of code related to drawing.

This makes DEBUG less verbose and more useful for the low level protocol stuff
it was intended(?) for.