wfreerdp: use LowLevelKeyboardProc hook for keyboard input
It seems like this is better for our purpose than processing ordinary events.

This also makes it possible to intercept Alt+Tab and process it remotely.

The low level events are not tied to a window or to having focus, so we have to
maintain a global variable telling us which window have focus.