wfreerdp: use char** __argv for command line parsing instead of multibyte crazyness
We must handle unicode correctly inside the application, but "correct" handling
of unicode outside the codepage range on the command line is not needed - and
will probably never work from the command line anyway. It is better to keep the
command line parser simple and similar to the one in xfreerdp.

Let's take the overhead of custom support for windows the day we need it.