libfreerdpkbd: use libxkbfile directly instead of invoking xprop and setxkbmap
Calling directly is faster and removes some dependencies and is less fragile as
it avoids parsing the output.

This removes the secret indirect dependency to xkb and makes it explicit and optional.
Most users should now make sure they have xkbfile development headers installed
so they get the full functionality.

This also makes sure that the current xkb keyboard layout 'group' is used even
if it isn't the primary group.

Note that display now has to be initialized before we can call xf_kb_init, so
the keyboard initialization is shuffled around a bit.