xfreerdp: Fix Pause/Break handling
It is hard to say what the Pause key really _is_, but combined with Ctrl (as
Break) it has to be sent extended, and without Ctrl is something else no matter
what. It is thus most correct to map Pause to extended 0x46=70.

Pressing Pause without Ctrl is a one-shot event sent as a Ctrl+NumLock. It
seems like this broke long time ago, I recently removed the broken code, and
now it is introduced again in a different and apparently correctly working way.

On Windows the Pause event is un-extended 0x45=69, not to be confused with
NumLock.