initialize streams to have .end at the requested size
A new stream is obviusly empty and can only be used for writing. After
something has been written it is necessary (or recommended) to move .end anyway
before it is used for some kind of read.

For a new write buffer it is more relevant to know how much can be written than
to know that nothing has been written yet.

This makes the debug stream assertions work again.